Move mock to own file and update usage
