+++ firefox-ios/Client/Frontend/Browser/BrowserViewController/BrowserViewControllerLayoutManager.swift
@@ -65,7 +65,7 @@ final class BrowserViewControllerLayoutManager {
         readerModeBar.heightAnchor.constraint(equalToConstant: UIConstants.ToolbarHeight).isActive = true
     }
 
-    // MARK - Private helpers
+    // MARK: - Private helpers
 
     private func updateHeaderHeightConstraint(isBottomSearchBar: Bool) {
         guard isBottomSearchBar else {
